Add accessibility UI APIs and fix bridge for bound property values
Fix the Noesis Lua bridge to return cached binding results instead of
nil for data-bound properties (NsHelpers.inl StoredValue expression
handling).  The previous code was a TODO stub that pushed nil for all
expression-type StoredValues.  ComplexValue::base already holds the
evaluated result, so we read it the same as non-expression complex
values.  This makes GetProperty work for bound TextBlock.Text,
ContentPresenter.Content, and all other data-bound properties.

Add new accessibility APIs to Ext.UI:
- GetDataContext: reads inherited DataContext via DependencyProperty
  GetValue, bypassing the bridge StoredValue path (inherited DPs have
  no StoredValue on the child element)
- FindNameInWidget: cross-NameScope named element lookup via C++
- GetTopmostWidget: returns topmost visible UIWidget
- IsHitTestVisible, IsMoveFocusFocusable, HasLocalValue: element queries
- SubscribeDPChanged/UnsubscribeDPChanged: DP change subscription

Expand GlobalFocusMonitor with:
- Widget set change detection (Strategy 4) for overlay/dialog detection
- IsVisible pruning for efficient tree walks
- ls:MoveFocus.IsFocused support (Larian controller focus system)
- Lazy FocusedElement property discovery from widget mValues
- Widget settle delay to avoid spurious fires during UI rebuilds

Add Tolk.IsSpeaking for speech queue drain timing.
Fix signed/unsigned comparison warning in Enumerations.cpp.

+++ BG3Extender/GameDefinitions/Enumerations.cpp
@@ -81,7 +81,7 @@ void RegisterEnumerationTypeInformation()
     for (auto i = 0; i < std::size(gEnumRegistrationTable); i++) {
         auto const& entry = gEnumRegistrationTable[i];
         if (entry.Type == EnumRegistrationType::EnumType) {
-            if (typeInfos.size() <= entry.TypeId) {
+            if (typeInfos.size() <= static_cast<size_t>(entry.TypeId)) {
                 typeInfos.resize(entry.TypeId + 1);
             }
